Remove finally clause
I ran into an unexpected undefined variable error inside the
finally clause. It's not worth figuring it out; the SMTP
connections aren't guaranteed to be long-term so the
server on the other side of the connection likely won't
be bothered by a missing quit signal. Will figure this out
later.